Properly build environment to call pg_dump
Instead of building a command-line that might be subject to
a row of issues (the command-line used before carried the
database password in plaintext, making it available to everyone
with read access to the system, and was prone to shell injection
and syntax breakage if any setting contained special characters),
we now build a copy of the environment with all the libpq variables
that are needed prefilled. The password is written into a .pgpass
file as suggested by the PostgreSQL docs.

In order to make all that re-usable and easy to read and
understand, I decided to pack the environment generation
and .pgpass file handling in a context manager.

This also makes the call more compatible with various setups. For
instance, connecting to 127.0.0.1:5432 via TCP is something entirely
different than leaving the host empty — the latter causes pg_dump
to connect throug hthe UNIX socket in /var/run, which as completely
different semantics like different authentication methods, etc.

Closes #2.